+++ src/Home/index.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import { BrowserRouter as Router } from 'react-router-dom';
-import Home from './Home';
+import Home from './index';
 
 test('renders the route menu', () => {
   render(<Router><Home /></Router>);
